+++ app/ide-desktop/lib/dashboard/src/components/JSONSchemaInput.tsx
@@ -9,55 +9,6 @@ import Dropdown from '#/components/Dropdown'
 import * as jsonSchema from '#/utilities/jsonSchema'
 import * as object from '#/utilities/object'
 
-// =====================
-// === getSchemaName ===
-// =====================
-
-const SCHEMA_NAMES = new WeakMap<object, string>()
-
-/** Return a human-readable name representing a schema. */
-function getSchemaNameHelper(defs: Record<string, object>, schema: object): string {
-  if ('title' in schema) {
-    return String(schema.title)
-  } else if ('type' in schema) {
-    return String(schema.type)
-  } else if ('$ref' in schema) {
-    const referencedSchema = jsonSchema.lookupDef(defs, schema)
-    return referencedSchema == null ? '(unknown)' : getSchemaName(defs, referencedSchema)
-  } else if ('anyOf' in schema) {
-    const members = Array.isArray(schema.anyOf) ? schema.anyOf : []
-    return (
-      members
-        .flatMap(object.singletonObjectOrNull)
-        .map(childSchema => getSchemaName(defs, childSchema))
-        .join(' | ') || '(unknown)'
-    )
-  } else if ('allOf' in schema) {
-    const members = Array.isArray(schema.allOf) ? schema.allOf : []
-    return (
-      members
-        .flatMap(object.singletonObjectOrNull)
-        .map(childSchema => getSchemaName(defs, childSchema))
-        .join(' & ') || '(unknown)'
-    )
-  } else {
-    return '(unknown)'
-  }
-}
-
-/** Return a human-readable name representing a schema.
- * This function is a memoized version of {@link getSchemaNameHelper}. */
-function getSchemaName(defs: Record<string, object>, schema: object) {
-  const cached = SCHEMA_NAMES.get(schema)
-  if (cached != null) {
-    return cached
-  } else {
-    const name = getSchemaNameHelper(defs, schema)
-    SCHEMA_NAMES.set(schema, name)
-    return name
-  }
-}
-
 // =======================
 // === JSONSchemaInput ===
 // =======================
@@ -341,7 +292,7 @@ export default function JSONSchemaInput(props: JSONSchemaInputProps) {
           readOnly={readOnly}
           items={childSchemas}
           selectedIndex={selectedChildIndex}
-          render={childProps => getSchemaName(defs, childProps.item)}
+          render={childProps => jsonSchema.getSchemaName(defs, childProps.item)}
           className="self-start"
           onClick={(childSchema, index) => {
             setSelectedChildIndex(index)
